+++ elastic-job-lite-core/src/main/resources/META-INF/services/org.apache.shardingsphere.elasticjob.lite.handler.threadpool.JobExecutorServiceHandler
@@ -15,6 +15,5 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.elasticjob.lite.executor.handler.error.impl.LogJobErrorHandler
-org.apache.shardingsphere.elasticjob.lite.executor.handler.error.impl.IgnoreJobErrorHandler
-org.apache.shardingsphere.elasticjob.lite.executor.handler.error.impl.ThrowJobErrorHandler
+org.apache.shardingsphere.elasticjob.lite.handler.threadpool.impl.CPUUsageJobExecutorServiceHandler
+org.apache.shardingsphere.elasticjob.lite.handler.threadpool.impl.SingleThreadJobExecutorServiceHandler
